pytorch/requirements.txt
# -*- coding: utf-8 -*-
torch
pytorch-ignite
numpy
